Frontend StudentContext
