+++ frontend/src/context/index.ts
@@ -1,0 +1,2 @@
+export * from './CourseContext'
+export * from './StudentContext'
