--- frontend/src/context/index.tsx
@@ -1,1 +1,0 @@
-export * from './CourseContext'
