+++ frontend/src/modules/Core/Types/index.ts
@@ -4,4 +4,5 @@ export * from '@core/Types/Questions'
 export * from '@core/Types/CustomComponent'
 export * from '@core/Types/EmailTemplates'
 export * from '@core/Types/Course'
+export * from '@core/Types/Student'
 export * from '@core/Components'
