+++ frontend/src/modules/EditZing/Types/Student.ts
@@ -1,34 +1,4 @@
-export interface Student {
-  name: string
-  email: string
-  year: string
-  college: string
-  groups: GroupMembership[]
-}
-
-export interface GroupMembership {
-  courseId: string
-  groupNumber: number
-  notes: string
-  notesModifyTime: Date
-  submissionTime: Date
-}
-
-export interface ResponseStudent {
-  name: string
-  email: string
-  year: string
-  college: string
-  groups: ResponseGroupMembership[]
-}
-
-export interface ResponseGroupMembership {
-  courseId: string
-  groupNumber: number
-  notes: string
-  notesModifyTime: string
-  submissionTime: string
-}
+import { Student } from '@core/Types'
 
 /** item type for drag and drop prop transfer via dnd */
 export type DnDStudentTransferType = {
@@ -41,15 +11,3 @@ export type DnDStudentTransferType = {
  * this case, this is a movable student type
  */
 export const STUDENT_TYPE = 'Student'
-
-export const responseStudentToStudent = (
-  student: ResponseStudent
-): Student => ({
-  ...student,
-  groups: student.groups.map((groupMembership) => ({
-    ...groupMembership,
-    // Remember to convert the notesModifyTime too
-    notesModifyTime: new Date(groupMembership.notesModifyTime),
-    submissionTime: new Date(groupMembership.submissionTime),
-  })),
-})
